requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
Werkzeug==3.0.1
email-validator==2.1.0.post1
requests==2.31.0
pandas==2.1.1
plotly==5.17.0
tweepy==4.14.0
facebook-sdk==3.1.0
python-linkedin-v2==0.9.4
instaloader==4.10.0
oauthlib==3.2.2
praw
vaderSentiment
matplotlib
seaborn
numpy
Flask-Mail 